multimodale_app/requirements.txt
flask
requests
python-dotenv
opencv-python-headless
werkzeug
